# api-docs.oas.yaml
openapi: 3.0.3
info:
  title: Cause connect API
  description: API for the cause connect service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  # Added by API Auto Mocking Plugin
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/Trevodorax/cause-connect/1.0.0
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/Trevodorax/TeaMaker/1.0.0
  - url: http://localhost:3000/
...
paths:
  /associations:
    get:
      tags:
        - association
      description: Get all associations
      responses:
        '200':
          description: Successfully retrieved associations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Association'
    post:
      tags:
        - association
      description: Create a new association
      requestBody:
        description: Informations about the new association
        content:
          application/json:
            schema:
              type: object
              required:
                - user
                - association
              properties:
                user:
                  $ref: '#/components/schemas/NewUser'
                association:
                  $ref: '#/components/schemas/NewAssociation'
        required: true
      responses:
        '200':
          description: Successfully created association
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Association'
        '422':
          description: Invalid association data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /associations/{associationId}:
    delete:
      tags:
        - association
      description: Delete an existing association
      parameters:
        - in: path
          name: associationId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully deleted association
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Association'
        '404':
          description: Association not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
        - association
      description: Get infos of an existing association
      parameters:
        - in: path
          name: associationId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved association
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Association'
        '404':
          description: Association not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags: 
        - association
      description: Change infos of an existing association
      parameters:
        - in: path
          name: associationId
          required: true
          schema:
            type: string
      requestBody:
        description: Informations to update
        content:
          application/json:
            schema:
              type: object
              required:
                - association
              properties:
                association:
                  $ref: '#/components/schemas/PartialAssociation'
      responses:
        '200':
          description: Successfully updated association
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Association'
        '404':
          description: Association not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Invalid association data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /users:
    post:
      tags:
        - user
      description: Create a new user
      requestBody:
        description: Informations about the new user
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                fullName:
                  type: string
                role:
                  type: string
                  enum: [admin, internal, external]
        required: true
      responses:
        '200':
          description: Successfully created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Invalid user data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
        - user
      description: Get all users from the association of the authenticated user
      parameters:
        - in: query
          name: role
          schema:
            type: string
            enum: [admin, internal, external]
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /users/{userId}:
    patch:
      tags:
        - user
      description: Update an existing user
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        description: Informations to update
        content:
          application/json:
            schema:
              type: object
              required:
                - user
              properties:
                user:
                  $ref: '#/components/schemas/PartialUser'
      responses:
        '200':
          description: Successfully updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - user
      description: Delete an existing user
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully deleted user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userId}/send-password-email:
    post:
      tags:
        - user
      description: Send an email to the user to reset his password
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully sent email
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    NewUser:
      type: object
      required:
        - email
        - fullName
      properties:
        email:
          type: string
        fullName:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [admin, internal, external]
    PartialUser:
      type: object
      properties:
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [admin, internal, external]
    NewAssociation:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
        description:
          type: string
        logo:
          type: string
          format: base64
    PartialAssociation:
      type: object
      properties:
        name:
          type: string
        logo:
          type: string
        description:
          type: string
    Association:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        logo:
          type: string
        description:
          type: string
...
    Error:
      required:
        - message
      type: object
      properties:
        message:
          type: string